+++ server-kit/src/persistence/WorkspaceManager.js
@@ -3,6 +3,8 @@ import { Backend } from '../api/Backend.js';
 import { SettingsManager } from './SettingsManager.js';
 import { StorageAdapter } from './StorageAdapter.js';
 import { ModalSystem } from '../ui/ModalSystem.js';
+import { checkSplashMessage } from '../ui/SplashMessage.js';
+import { GeolocationManager } from '../core/geospatial/GeolocationManager.js';
 
 export class WorkspaceManager {
 	static context = null;
@@ -16,6 +18,8 @@ export class WorkspaceManager {
 		this.context.AppState.workspace.isInitializing = true;
 
 		await Security.init();
+
+		checkSplashMessage().then(() => GeolocationManager.setupGeolocation());
 
 		this.context.AppState.setSaveCallback(this.saveWorkspaceSettings.bind(this));